decompose UserPostsContexts into smaller hooks
